Update Device Bidding
Only look at line 1-17. The following variables require the SEM's attention. 1-6 should be self explanatory. Add text that is in all campaigns you want looked at (do not use dashes; that will cause errors. underscores will work though. in highConvMaxCplDsk, highConvMaxCplMob, and highConvMaxCplTab enter the highest cpl you would be willing to pay for consistent conversions for each device type. The script is designed to be run once daily using micro bid changes looking at if-then cpl's in a series of time ranges. Over time, it should achieve an equilibrium, especially in concert with other bidding scripts. 'usePreviousDayData' evaluates if the performance from the day before warranted changes. It requires the CPL be much higher or much lower than the other time frames and is partially in place in case a particular day has a large swing that may need a correction. 'useAllTimeData' is not used for bidding. It is used to avoid the script over-correcting against the historical performance. For example, this portion reduces the bid change if the devices with a long term good cpl have short term negative swings, and bid increases are reduced on devices with great performance short term and bad performance long term. This is the only purpose of the date at the bottom. The format must remain as listed. If an account has had a lot of rebuilding or there were major conversion issues, it is best to set the date some time after that time period so that it is using timely data. 'useVariedDays' uses as an algorithm to determine whether it should apply each bid change. Stronger performance anomalies are weighted more heavily by recent changes. This helps avoid over-correcting and helps reach the equilibrium and stay within that range. The last variable is 'clearAdGroups'. Since the script bids only at the campaign level, and ad group level device bid modifiers take priority over campaign level, this setting looks for ad groups with device modifiers and clears them. The script will not perform at the optimal level otherwise. All calculations are factored to avoid bidding twice on identical data in the same time period. To explain further, a brand new campaign with 6 days of data would fall within the 7 days, 14 days, and 30 days. It is only once there is data from 15 days that the 14 days kicks in, and it is only after 31 days of data that the 30 days kicks in. Also devices without an impression from the day before are not evaluated either as they were assumed to be turned off. Last note. Any of the variables can be turned off by changing the value to false. If the SEM wants to run the script weekly instead of daily, it is best to turn off 'useVariedDays' and 'usePreviousDayData', but that is the only difference.

var campContains1 = 'EnterPortionOfCampaignName';
var campContains2 = 'EnterPortionOfCampaignName';
var campContains3 = 'EnterPortionOfCampaignName';
var campDoesNotContain1 = 'Placeholder';
var campDoesNotContain2 = 'Placeholder';
var campDoesNotContain3 = 'Placeholder';

var highConvMaxCplDsk = 165;
var highConvMaxCplMob = 165;
var highConvMaxCplTab = 165;
var usePreviousDayData = true;
var useAllTimeData = true;
var useVariedDays = true;
var clearAdGroups = true;

var start=new Date("January 01, 2017 00:00:00");