Create Display Exclusions
The setup on this script can take a little time, but the power makes it worth it. There are a number of variables to look at. First, the script is designed to look for url's and domains in the desired campaigns to add placement exclusions. It was intended to be used on Campaign Exclusion Lists, which is what the first line is for. Each list name needs to be in apostrophe's with a comma separating additional ones. You will need to create the lists first and then attach them to campaigns. If you don't want to use lists, there are options available to skip creating them. The second row uses numbers with no apostrophe's. This should be the lowest cpl number that you would immediately add a negative for. Each value should correspond to the the one above it. The value in the second place is for the second list and so on. Then there are a series of values to have the script look at campaigns that include certain text and then ones that they don't include. Certain characters like dashes cause the script errors (but underscores are ok). Each value in order applies to the value in order for the variables above it. The next variable is the date. The script will look from that date through yesterday for its evaluations. If the account or many campaigns have been rebuilt recently, it's best not to set the date before that time. But the format must remain as-is. Then next variable is 'autoAdd' which is set to true but can be turned off by setting the value to false. If it is set to true, the script will automatically add the placement exclusion to list or campaign. The variable 'printToSheet' is for those that want to have the placement exclusions listed on a sheet and then emailed to them. autoAdd and printToSheet can both run so the SEM can allow the script to add the exclusions and the SEM sees what was evaluated. 'listExclusions is intended if you're using placement exclusion lists. If set to false, simply add a dummy list name in 'listNames'. 'campExclusions' is for adding the exclusions to campaigns. I will describe 'lowConverting' further down below. Next, if a sheet is to be used, create a google sheet and include the url as listed below. Added the email address the sheet will be sent to and the account name that will appear in the subject. At the bottom of the Spreadsheet, name the Sheet(tab) and include that name where it says ENTER_SHEET_NAME.

The script has a simple evaluation. It looks at urls and domains and uses the 'threshold' value. If the CPL or the spend without conversions is equal to or above that it adds the exclusion. When using lists, it does not aggregate all the placement instances for the evaluated campaigns to find the data. It simply looks through each of them to add. So if there are very different campaigns that might serve to the same placement, it would be better to have separate lists or use the campExclusions option instead. The 'lowConverting' variable evaluates placements with spend that is only 60% the threshold value but have at least 100 or more clicks and less than 1% conversion rate. If the cpl is under half the threshold value, the query is not added as an exclusion. This evaluation is intended to capture placements that have a slow burn and low spend and doesn't capture very many. It's more of a safeguard against those placements that are rarely a problem at any one point in time.

var listNames=['List 1', 'List 2'];
var threshold=[600, 600];
var campContains1=['CampaignText', 'CampaignText'];
var campContains2=['CampaignText', 'CampaignText'];
var campContains3=['CampaignText', 'CampaignText'];
var campDoesNotContain1=['Placeholder', 'Placeholder'];
var campDoesNotContain2=['Placeholder', 'Placeholder'];
var campDoesNotContain3=['Placeholder', 'Placeholder'];

var starting=new Date("January 01, 2017 00:00:00");

var autoAdd = true;
var printToSheet = true;
var listExclusions = true;
var campExclusions = true;
var lowConverting = true;

var emailaddress='INSERT_EMAIL_ADDRESS';
var accountname='INSERT_ACCOUNT_NAME';
var URL='ENTER_SPREADSHEET_URL';

var sheet=spreadsheet.getSheetByName('ENTER_SHEET_NAME');