Create Ad Schedule Bids
Only look at line 1-12. The following variables require the SEM's attention. 1-6 should be self explanatory. Add text that is in all campaigns you want looked at (do not use dashes; that will cause errors. underscores will work though. in highConvMaxCpl, enter the highest cpl you would be willing to pay for consistent conversions in a keyword. The script is designed to be run once weekly using micro bid changes looking at if-then cpl's in a series of time ranges. It looks only at the ad schedule segments for each campaign. So if no ad schedule is set, it won't do anything. If an ad schedule is set for 100% of the time but not segmented, it will set a bid modifier for each entire day. Over time, it should achieve an equilibrium, especially in concert with other bidding scripts. 'useAllTimeData' is not used for bidding. It is used to avoid the script over-correcting against the historical performance. For example, this portion reduces the bid change if certain times with a long term good cpl have short term negative swings, and bid increases are reduced on times with great performance short term and bad performance long term. This is the only purpose of the date at the bottom. The format must remain as listed. If a an account has had a lot of rebuilding or there were major conversion issues, it is best to set the date some time after that time period so that it is using timely data. 'useVariedDays' uses as an algorithm to determine whether it should apply each bid change. Stronger performance anomalies are weighted more heavily by recent changes. This helps avoid over-correcting and helps reach the equilibrium and stay within that range. All calculations are factored to avoid bidding twice on identical data in the same time period. To explain further, a brand new campaign with 6 days of data would fall within the 7 days, 14 days, and 30 days. It is only once there is data from 15 days that the 14 days kicks in, and it is only after 31 days of data that the 30 days kicks in. Also time segments without an impression from the week prior are not evaluated either as they were assumed to be turned off. Last note. Any of the variables can be turned off by changing the value to false.