Create Demographic Bids
Only look at line 1-20. The following variables require the SEM's attention. 1-6 should be self explanatory. Add text that is in all campaigns you want looked at (do not use dashes; that will cause errors. underscores will work though). in highConvMaxCpl, enter the highest cpl you would be willing to pay for consistent conversions. The script is designed to be run once weekly using micro bid changes looking at if-then cpl's in a series of biddable demographics. It looks only at the demographics for each adgroup. While demographic data is automatically collected by AdWords, if the demographic is not set as a biddable demographic (for example, added into the ad groups in Editor), the bids will not be set. Over time, it should achieve an equilibrium, especially in concert with other bidding scripts. 'useAllTimeData' is not used for bidding. It is used to avoid the script over-correcting against the historical performance. For example, this portion reduces the bid change if certain demographics with a long term good cpl have short term negative swings, and bid increases are reduced on demographics with great performance short term and bad performance long term. This is the only purpose of the date at the bottom. The format must remain as listed. If an account has had a lot of rebuilding or there were major conversion issues, it is best to set the date some time after that time period so that it is using timely data. 'useVariedDays' uses as an algorithm to determine whether it should apply each bid change. Stronger performance anomalies are weighted more heavily by recent changes. This helps avoid over-correcting and helps reach the equilibrium and stay within that range. All calculations are factored to avoid bidding twice on identical data in the same time period. To explain further, a brand new campaign with 6 days of data would fall within the 7 days, 14 days, and 30 days. It is only once there is data from 15 days that the 14 days kicks in, and it is only after 31 days of data that the 30 days kicks in. Also demographics without an impression from the week prior are not evaluated either as they were assumed to be turned off. Any of the variables can be turned off by changing the value to false. Scripts cannot currently modify demographic bids. Thus, the script is designed to log data on a Google Sheet pre-formatted to be able to upload to AdWords Editor.