Create Display AdGroup Bids
Only look at line 1-13. The following variables require the SEM's attention. 1-6 should be self explanatory. Include text that is in all campaigns you want looked at (do not use dashes; that will cause errors. underscores will work though. in highConvMaxCpl, enter the highest cpl you would be willing to pay for consistent conversions.. The script is designed to be run once daily using micro bid changes looking at if-then cpl's in a series of time ranges. Over time, it should achieve an equilibrium, especially in concert with other bidding scripts. The script discovers what targeting type each display ad group uses for a primary bid. As long as it is either 'none', which uses the ad group's bid itself, keyword, topic, audience or placement, it calculates the performance and changes that bid.

'usePreviousDayData' evaluates if the performance from the day before warranted changes. It requires the CPL be much higher or much lower than the other time frames and is partially in place in case a particular day has a large swing that may need a correction. 'useAllTimeData' is not used for bidding. It is used to avoid the script over-correcting against the historical performance. For example, this portion reduces the bid change if the display targeting with long term good cpl has short term negative swings, and bid increases are reduced on display targeting with great performance short term and bad performance long term. This is the only purpose of the date at the bottom. The format must remain as listed. If a an account has had a lot of rebuilding or there were major conversion issues, it is best to set the date some time after that time period so that it is using timely data. 'useVariedDays' uses as an algorithm to determine whether it should apply each bid change. Stronger performance anomalies are weighted more heavily by recent changes. This helps avoid over-correcting and helps reach the equilibrium and stay within that range. All calculations are factored to avoid bidding twice on identical data in the same time period. To explain further, a brand new campaign with 6 days of data would fall within the 7 days, 14 days, and 30 days. It is only once there is data from 15 days that the 14 days kicks in, and it is only after 31 days of data that the 30 days kicks in. Also display targeting without an impression from the day before are not evaluated either as they were assumed to be turned off. Last note. Any of the variables can be turned off by changing the value to false. If the SEM wants to run the script weekly instead of daily, it is best to turn off 'useVariedDays' and 'usePreviousDayData', but that is the only difference.